Change box dimensions to unsigned ints

Closes #8

See merge request gitschooldude/hello!17

+++ include/box.hh
@@ -3,9 +3,9 @@ class Box
    public:
       Box();
       std::string name;
-      double length;   // Length of a box
-      double width;    // width of a box
-      double height;   // Height of a box
+      unsigned int length;   // Length of a box
+      unsigned int width;    // width of a box
+      unsigned int height;   // Height of a box
 
       void report();
       void setWidth( double inWidth) ;
